Model Functions

gmsh.model.add — Function

gmsh.model.add(name)

Add a new model, with name name, and set it as the current model.

gmsh.model.remove — Function

gmsh.model.remove()

Remove the current model.

gmsh.model.list — Function

gmsh.model.list()

List the names of all models.

Return names.

gmsh.model.getCurrent — Function

gmsh.model.getCurrent()

Get the name of the current model.

Return name.

gmsh.model.setCurrent — Function

gmsh.model.setCurrent(name)

Set the current model to the model with name name. If several models have the same name, select the one that was added first.

gmsh.model.getEntities — Function

gmsh.model.getEntities(dim = -1)

Get all the entities in the current model. If dim is >= 0, return only the entities of the specified dimension (e.g. points if dim == 0). The entities are returned as a vector of (dim, tag) integer pairs.

Return dimTags.

gmsh.model.setEntityName — Function

gmsh.model.setEntityName(dim, tag, name)

Set the name of the entity of dimension dim and tag tag.

gmsh.model.getEntityName — Function

gmsh.model.getEntityName(dim, tag)

Get the name of the entity of dimension dim and tag tag.

Return name.

gmsh.model.getPhysicalGroups — Function

gmsh.model.getPhysicalGroups(dim = -1)

Get all the physical groups in the current model. If dim is >= 0, return only the entities of the specified dimension (e.g. physical points if dim == 0). The entities are returned as a vector of (dim, tag) integer pairs.

Return dimTags.

gmsh.model.getEntitiesForPhysicalGroup — Function

gmsh.model.getEntitiesForPhysicalGroup(dim, tag)

Get the tags of the model entities making up the physical group of dimension dim and tag tag.

Return tags.

gmsh.model.getPhysicalGroupsForEntity — Function

gmsh.model.getPhysicalGroupsForEntity(dim, tag)

Get the tags of the physical groups (if any) to which the model entity of dimension dim and tag tag belongs.

Return physicalTags.

gmsh.model.addPhysicalGroup — Function

gmsh.model.addPhysicalGroup(dim, tags, tag = -1)

Add a physical group of dimension dim, grouping the model entities with tags tags. Return the tag of the physical group, equal to tag if tag is positive, or a new tag if tag < 0.

Return an integer value.

gmsh.model.removePhysicalGroups — Function

gmsh.model.removePhysicalGroups(dimTags = Tuple{Cint,Cint}[])

Remove the physical groups dimTags from the current model. If dimTags is empty, remove all groups.

gmsh.model.setPhysicalName — Function

gmsh.model.setPhysicalName(dim, tag, name)

Set the name of the physical group of dimension dim and tag tag.

gmsh.model.removePhysicalName — Function

gmsh.model.removePhysicalName(name)

Remove the physical name name from the current model.

gmsh.model.getPhysicalName — Function

gmsh.model.getPhysicalName(dim, tag)

Get the name of the physical group of dimension dim and tag tag.

Return name.

gmsh.model.getBoundary — Function

gmsh.model.getBoundary(dimTags, combined = true, oriented = true, recursive = false)

Get the boundary of the model entities dimTags. Return in outDimTags the boundary of the individual entities (if combined is false) or the boundary of the combined geometrical shape formed by all input entities (if combined is true). Return tags multiplied by the sign of the boundary entity if oriented is true. Apply the boundary operator recursively down to dimension 0 (i.e. to points) if recursive is true.

Return outDimTags.

gmsh.model.getAdjacencies — Function

gmsh.model.getAdjacencies(dim, tag)

Get the upward and downward adjacencies of the model entity of dimension dim and tag tag. The upward vector returns the adjacent entities of dimension dim + 1; the downward vector returns the adjacent entities of dimension dim - 1.

Return upward, downward.

gmsh.model.getEntitiesInBoundingBox — Function

gmsh.model.getEntitiesInBoundingBox(xmin, ymin, zmin, xmax, ymax, zmax, dim = -1)

Get the model entities in the bounding box defined by the two points (xmin, ymin, zmin) and (xmax, ymax, zmax). If dim is >= 0, return only the entities of the specified dimension (e.g. points if dim == 0).

Return tags.

gmsh.model.getBoundingBox — Function

gmsh.model.getBoundingBox(dim, tag)

Get the bounding box (xmin, ymin, zmin), (xmax, ymax, zmax) of the model entity of dimension dim and tag tag. If dim and tag are negative, get the bounding box of the whole model.

Return xmin, ymin, zmin, xmax, ymax, zmax.

gmsh.model.getDimension — Function

gmsh.model.getDimension()

Get the geometrical dimension of the current model.

Return an integer value.

gmsh.model.addDiscreteEntity — Function

gmsh.model.addDiscreteEntity(dim, tag = -1, boundary = Cint[])

Add a discrete model entity (defined by a mesh) of dimension dim in the current model. Return the tag of the new discrete entity, equal to tag if tag is positive, or a new tag if tag < 0. boundary specifies the tags of the entities on the boundary of the discrete entity, if any. Specifying boundary allows Gmsh to construct the topology of the overall model.

Return an integer value.

gmsh.model.removeEntities — Function

gmsh.model.removeEntities(dimTags, recursive = false)

Remove the entities dimTags of the current model. If recursive is true, remove all the entities on their boundaries, down to dimension 0.

gmsh.model.removeEntityName — Function

gmsh.model.removeEntityName(name)

Remove the entity name name from the current model.

gmsh.model.getType — Function

gmsh.model.getType(dim, tag)

Get the type of the entity of dimension dim and tag tag.

Return entityType.

gmsh.model.getParent — Function

gmsh.model.getParent(dim, tag)

In a partitioned model, get the parent of the entity of dimension dim and tag tag, i.e. from which the entity is a part of, if any. parentDim and parentTag are set to -1 if the entity has no parent.

Return parentDim, parentTag.

gmsh.model.getPartitions — Function

gmsh.model.getPartitions(dim, tag)

In a partitioned model, return the tags of the partition(s) to which the entity belongs.

Return partitions.

gmsh.model.getValue — Function

gmsh.model.getValue(dim, tag, parametricCoord)

Evaluate the parametrization of the entity of dimension dim and tag tag at the parametric coordinates parametricCoord. Only valid for dim equal to 0 (with empty parametricCoord), 1 (with parametricCoord containing parametric coordinates on the curve) or 2 (with parametricCoord containing pairs of u, v parametric coordinates on the surface, concatenated: [p1u, p1v, p2u, ...]). Return triplets of x, y, z coordinates in coord, concatenated: [p1x, p1y, p1z, p2x, ...].

Return coord.

gmsh.model.getDerivative — Function

gmsh.model.getDerivative(dim, tag, parametricCoord)

Evaluate the derivative of the parametrization of the entity of dimension dim and tag tag at the parametric coordinates parametricCoord. Only valid for dim equal to 1 (with parametricCoord containing parametric coordinates on the curve) or 2 (with parametricCoord containing pairs of u, v parametric coordinates on the surface, concatenated: [p1u, p1v, p2u, ...]). For dim equal to 1 return the x, y, z components of the derivative with respect to u [d1ux, d1uy, d1uz, d2ux, ...]; for dim equal to 2 return the x, y, z components of the derivative with respect to u and v: [d1ux, d1uy, d1uz, d1vx, d1vy, d1vz, d2ux, ...].

Return derivatives.

gmsh.model.getSecondDerivative — Function

gmsh.model.getSecondDerivative(dim, tag, parametricCoord)

Evaluate the second derivative of the parametrization of the entity of dimension dim and tag tag at the parametric coordinates parametricCoord. Only valid for dim equal to 1 (with parametricCoord containing parametric coordinates on the curve) or 2 (with parametricCoord containing pairs of u, v parametric coordinates on the surface, concatenated: [p1u, p1v, p2u, ...]). For dim equal to 1 return the x, y, z components of the second derivative with respect to u [d1uux, d1uuy, d1uuz, d2uux, ...]; for dim equal to 2 return the x, y, z components of the second derivative with respect to u and v, and the mixed derivative with respect to u and v: [d1uux, d1uuy, d1uuz, d1vvx, d1vvy, d1vvz, d1uvx, d1uvy, d1uvz, d2uux, ...].

Return derivatives.

gmsh.model.getCurvature — Function

gmsh.model.getCurvature(dim, tag, parametricCoord)

Evaluate the (maximum) curvature of the entity of dimension dim and tag tag at the parametric coordinates parametricCoord. Only valid for dim equal to 1 (with parametricCoord containing parametric coordinates on the curve) or 2 (with parametricCoord containing pairs of u, v parametric coordinates on the surface, concatenated: [p1u, p1v, p2u, ...]).

Return curvatures.

gmsh.model.getPrincipalCurvatures — Function

gmsh.model.getPrincipalCurvatures(tag, parametricCoord)

Evaluate the principal curvatures of the surface with tag tag at the parametric coordinates parametricCoord, as well as their respective directions. parametricCoord are given by pair of u and v coordinates, concatenated: [p1u, p1v, p2u, ...].

Return curvatureMax, curvatureMin, directionMax, directionMin.

gmsh.model.getNormal — Function

gmsh.model.getNormal(tag, parametricCoord)

Get the normal to the surface with tag tag at the parametric coordinates parametricCoord. parametricCoord are given by pairs of u and v coordinates, concatenated: [p1u, p1v, p2u, ...]. normals are returned as triplets of x, y, z components, concatenated: [n1x, n1y, n1z, n2x, ...].

Return normals.

gmsh.model.getParametrization — Function

gmsh.model.getParametrization(dim, tag, coord)

Get the parametric coordinates parametricCoord for the points coord on the entity of dimension dim and tag tag. coord are given as triplets of x, y, z coordinates, concatenated: [p1x, p1y, p1z, p2x, ...]. parametricCoord returns the parametric coordinates t on the curve (if dim = 1) or pairs of u and v coordinates concatenated on the surface (if dim = 2), i.e. [p1t, p2t, ...] or [p1u, p1v, p2u, ...].

Return parametricCoord.

gmsh.model.getParametrizationBounds — Function

gmsh.model.getParametrizationBounds(dim, tag)

Get the min and max bounds of the parametric coordinates for the entity of dimension dim and tag tag.

Return min, max.

gmsh.model.isInside — Function

gmsh.model.isInside(dim, tag, parametricCoord)

Check if the parametric coordinates provided in parametricCoord correspond to points inside the entitiy of dimension dim and tag tag, and return the number of points inside. This feature is only available for a subset of curves and surfaces, depending on the underyling geometrical representation.

Return an integer value.

gmsh.model.getClosestPoint — Function

gmsh.model.getClosestPoint(dim, tag, coord)

Get the points closestCoord on the entity of dimension dim and tag tag to the points coord, by orthogonal projection. coord and closestCoord are given as triplets of x, y, z coordinates, concatenated: [p1x, p1y, p1z, p2x, ...]. parametricCoord returns the parametric coordinates t on the curve (if dim = 1) or pairs of u and v coordinates concatenated on the surface (if dim = 2), i.e. [p1t, p2t, ...] or [p1u, p1v, p2u, ...].

Return closestCoord, parametricCoord.

gmsh.model.reparametrizeOnSurface — Function

gmsh.model.reparametrizeOnSurface(dim, tag, parametricCoord, surfaceTag, which = 0)

Reparametrize the boundary entity (point or curve, i.e. with dim == 0 or dim == 1) of tag tag on the surface surfaceTag. If dim == 1, reparametrize all the points corresponding to the parametric coordinates parametricCoord. Multiple matches in case of periodic surfaces can be selected with which. This feature is only available for a subset of entities, depending on the underyling geometrical representation.

Return surfaceParametricCoord.

gmsh.model.setVisibility — Function

gmsh.model.setVisibility(dimTags, value, recursive = false)

Set the visibility of the model entities dimTags to value. Apply the visibility setting recursively if recursive is true.

gmsh.model.getVisibility — Function

gmsh.model.getVisibility(dim, tag)

Get the visibility of the model entity of dimension dim and tag tag.

Return value.

gmsh.model.setVisibilityPerWindow — Function

gmsh.model.setVisibilityPerWindow(value, windowIndex = 0)

Set the global visibility of the model per window to value, where windowIndex identifies the window in the window list.

gmsh.model.setColor — Function

gmsh.model.setColor(dimTags, r, g, b, a = 255, recursive = false)

Set the color of the model entities dimTags to the RGBA value (r, g, b, a), where r, g, b and a should be integers between 0 and 255. Apply the color setting recursively if recursive is true.

gmsh.model.getColor — Function

gmsh.model.getColor(dim, tag)

Get the color of the model entity of dimension dim and tag tag.

Return r, g, b, a.

gmsh.model.setCoordinates — Function

gmsh.model.setCoordinates(tag, x, y, z)

Set the x, y, z coordinates of a geometrical point.