OpenCASCADE CAD Kernel

The OpenCASCADE (occ) kernel is currently not natively shipped with gmsh_jll.jl. Thus, all functions only work if a full gmsh build is installed locally.

gmsh.model.occ.addPoint — Function

gmsh.model.occ.addPoint(x, y, z, meshSize = 0., tag = -1)

Add a geometrical point in the OpenCASCADE CAD representation, at coordinates (x, y, z). If meshSize is > 0, add a meshing constraint at that point. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the point. (Note that the point will be added in the current model only after synchronize is called. This behavior holds for all the entities added in the occ module.)

Return an integer value.

gmsh.model.occ.addLine — Function

gmsh.model.occ.addLine(startTag, endTag, tag = -1)

Add a straight line segment in the OpenCASCADE CAD representation, between the two points with tags startTag and endTag. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the line.

Return an integer value.

gmsh.model.occ.addCircleArc — Function

gmsh.model.occ.addCircleArc(startTag, centerTag, endTag, tag = -1)

Add a circle arc in the OpenCASCADE CAD representation, between the two points with tags startTag and endTag, with center centerTag. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the circle arc.

Return an integer value.

gmsh.model.occ.addCircle — Function

gmsh.model.occ.addCircle(x, y, z, r, tag = -1, angle1 = 0., angle2 = 2*pi)

Add a circle of center (x, y, z) and radius r in the OpenCASCADE CAD representation. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. If angle1 and angle2 are specified, create a circle arc between the two angles. Return the tag of the circle.

Return an integer value.

gmsh.model.occ.addEllipseArc — Function

gmsh.model.occ.addEllipseArc(startTag, centerTag, majorTag, endTag, tag = -1)

Add an ellipse arc in the OpenCASCADE CAD representation, between the two points startTag and endTag, and with center centerTag and major axis point majorTag. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the ellipse arc. Note that OpenCASCADE does not allow creating ellipse arcs with the major radius smaller than the minor radius.

Return an integer value.

gmsh.model.occ.addEllipse — Function

gmsh.model.occ.addEllipse(x, y, z, r1, r2, tag = -1, angle1 = 0., angle2 = 2*pi)

Add an ellipse of center (x, y, z) and radii r1 and r2 along the x- and y-axes, respectively, in the OpenCASCADE CAD representation. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. If angle1 and angle2 are specified, create an ellipse arc between the two angles. Return the tag of the ellipse. Note that OpenCASCADE does not allow creating ellipses with the major radius (along the x-axis) smaller than or equal to the minor radius (along the y-axis): rotate the shape or use addCircle in such cases.

Return an integer value.

gmsh.model.occ.addSpline — Function

gmsh.model.occ.addSpline(pointTags, tag = -1)

Add a spline (C2 b-spline) curve in the OpenCASCADE CAD representation, going through the points pointTags. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Create a periodic curve if the first and last points are the same. Return the tag of the spline curve.

Return an integer value.

gmsh.model.occ.addBSpline — Function

gmsh.model.occ.addBSpline(pointTags, tag = -1, degree = 3, weights = Cdouble[], knots = Cdouble[], multiplicities = Cint[])

Add a b-spline curve of degree degree in the OpenCASCADE CAD representation, with pointTags control points. If weights, knots or multiplicities are not provided, default parameters are computed automatically. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Create a periodic curve if the first and last points are the same. Return the tag of the b-spline curve.

Return an integer value.

gmsh.model.occ.addBezier — Function

gmsh.model.occ.addBezier(pointTags, tag = -1)

Add a Bezier curve in the OpenCASCADE CAD representation, with pointTags control points. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the Bezier curve.

Return an integer value.

gmsh.model.occ.addWire — Function

gmsh.model.occ.addWire(curveTags, tag = -1, checkClosed = false)

Add a wire (open or closed) in the OpenCASCADE CAD representation, formed by the curves curveTags. Note that an OpenCASCADE wire can be made of curves that share geometrically identical (but topologically different) points. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the wire.

Return an integer value.

gmsh.model.occ.addCurveLoop — Function

gmsh.model.occ.addCurveLoop(curveTags, tag = -1)

Add a curve loop (a closed wire) in the OpenCASCADE CAD representation, formed by the curves curveTags. curveTags should contain tags of curves forming a closed loop. Note that an OpenCASCADE curve loop can be made of curves that share geometrically identical (but topologically different) points. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the curve loop.

Return an integer value.

gmsh.model.occ.addRectangle — Function

gmsh.model.occ.addRectangle(x, y, z, dx, dy, tag = -1, roundedRadius = 0.)

Add a rectangle in the OpenCASCADE CAD representation, with lower left corner at (x, y, z) and upper right corner at (x + dx, y + dy, z). If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Round the corners if roundedRadius is nonzero. Return the tag of the rectangle.

Return an integer value.

gmsh.model.occ.addDisk — Function

gmsh.model.occ.addDisk(xc, yc, zc, rx, ry, tag = -1)

Add a disk in the OpenCASCADE CAD representation, with center (xc, yc, zc) and radius rx along the x-axis and ry along the y-axis. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the disk.

Return an integer value.

gmsh.model.occ.addPlaneSurface — Function

gmsh.model.occ.addPlaneSurface(wireTags, tag = -1)

Add a plane surface in the OpenCASCADE CAD representation, defined by one or more curve loops (or closed wires) wireTags. The first curve loop defines the exterior contour; additional curve loop define holes. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface.

Return an integer value.

gmsh.model.occ.addSurfaceFilling — Function

gmsh.model.occ.addSurfaceFilling(wireTag, tag = -1, pointTags = Cint[])

Add a surface in the OpenCASCADE CAD representation, filling the curve loop wireTag. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface. If pointTags are provided, force the surface to pass through the given points.

Return an integer value.

gmsh.model.occ.addBSplineFilling — Function

gmsh.model.occ.addBSplineFilling(wireTag, tag = -1, type = "")

Add a BSpline surface in the OpenCASCADE CAD representation, filling the curve loop wireTag. The curve loop should be made of 2, 3 or 4 BSpline curves. The optional type argument specifies the type of filling: "Stretch" creates the flattest patch, "Curved" (the default) creates the most rounded patch, and "Coons" creates a rounded patch with less depth than "Curved". If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface.

Return an integer value.

gmsh.model.occ.addBezierFilling — Function

gmsh.model.occ.addBezierFilling(wireTag, tag = -1, type = "")

Add a Bezier surface in the OpenCASCADE CAD representation, filling the curve loop wireTag. The curve loop should be made of 2, 3 or 4 Bezier curves. The optional type argument specifies the type of filling: "Stretch" creates the flattest patch, "Curved" (the default) creates the most rounded patch, and "Coons" creates a rounded patch with less depth than "Curved". If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface.

Return an integer value.

gmsh.model.occ.addBSplineSurface — Function

gmsh.model.occ.addBSplineSurface(pointTags, numPointsU, tag = -1, degreeU = 3, degreeV = 3, weights = Cdouble[], knotsU = Cdouble[], knotsV = Cdouble[], multiplicitiesU = Cint[], multiplicitiesV = Cint[], wireTags = Cint[], wire3D = false)

Add a b-spline surface of degree degreeU x degreeV in the OpenCASCADE CAD representation, with pointTags control points given as a single vector [Pu1v1, ... PunumPointsUv1, Pu1v2, ...]. If weights, knotsU, knotsV, multiplicitiesU or multiplicitiesV are not provided, default parameters are computed automatically. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. If wireTags is provided, trim the b-spline patch using the provided wires: the first wire defines the external contour, the others define holes. If wire3D is set, consider wire curves as 3D curves and project them on the b-spline surface; otherwise consider the wire curves as defined in the parametric space of the surface. Return the tag of the b-spline surface.

Return an integer value.

gmsh.model.occ.addBezierSurface — Function

gmsh.model.occ.addBezierSurface(pointTags, numPointsU, tag = -1, wireTags = Cint[], wire3D = false)

Add a Bezier surface in the OpenCASCADE CAD representation, with pointTags control points given as a single vector [Pu1v1, ... PunumPointsUv1, Pu1v2, ...]. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. If wireTags is provided, trim the Bezier patch using the provided wires: the first wire defines the external contour, the others define holes. If wire3D is set, consider wire curves as 3D curves and project them on the Bezier surface; otherwise consider the wire curves as defined in the parametric space of the surface. Return the tag of the Bezier surface.

Return an integer value.

gmsh.model.occ.addTrimmedSurface — Function

gmsh.model.occ.addTrimmedSurface(surfaceTag, wireTags = Cint[], wire3D = false, tag = -1)

Trim the surface surfaceTag with the wires wireTags, replacing any existing trimming curves. The first wire defines the external contour, the others define holes. If wire3D is set, consider wire curves as 3D curves and project them on the surface; otherwise consider the wire curves as defined in the parametric space of the surface. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the trimmed surface.

Return an integer value.

gmsh.model.occ.addSurfaceLoop — Function

gmsh.model.occ.addSurfaceLoop(surfaceTags, tag = -1, sewing = false)

Add a surface loop (a closed shell) in the OpenCASCADE CAD representation, formed by surfaceTags. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the surface loop. Setting sewing allows to build a shell made of surfaces that share geometrically identical (but topologically different) curves.

Return an integer value.

gmsh.model.occ.addVolume — Function

gmsh.model.occ.addVolume(shellTags, tag = -1)

Add a volume (a region) in the OpenCASCADE CAD representation, defined by one or more surface loops shellTags. The first surface loop defines the exterior boundary; additional surface loop define holes. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the volume.

Return an integer value.

gmsh.model.occ.addSphere — Function

gmsh.model.occ.addSphere(xc, yc, zc, radius, tag = -1, angle1 = -pi/2, angle2 = pi/2, angle3 = 2*pi)

Add a sphere of center (xc, yc, zc) and radius r in the OpenCASCADE CAD representation. The optional angle1 and angle2 arguments define the polar angle opening (from -Pi/2 to Pi/2). The optional angle3 argument defines the azimuthal opening (from 0 to 2*Pi). If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the sphere.

Return an integer value.

gmsh.model.occ.addBox — Function

gmsh.model.occ.addBox(x, y, z, dx, dy, dz, tag = -1)

Add a parallelepipedic box in the OpenCASCADE CAD representation, defined by a point (x, y, z) and the extents along the x-, y- and z-axes. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the box.

Return an integer value.

gmsh.model.occ.addCylinder — Function

gmsh.model.occ.addCylinder(x, y, z, dx, dy, dz, r, tag = -1, angle = 2*pi)

Add a cylinder in the OpenCASCADE CAD representation, defined by the center (x, y, z) of its first circular face, the 3 components (dx, dy, dz) of the vector defining its axis and its radius r. The optional angle argument defines the angular opening (from 0 to 2*Pi). If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. Return the tag of the cylinder.

Return an integer value.

gmsh.model.occ.addCone — Function

gmsh.model.occ.addCone(x, y, z, dx, dy, dz, r1, r2, tag = -1, angle = 2*pi)

Add a cone in the OpenCASCADE CAD representation, defined by the center (x, y, z) of its first circular face, the 3 components of the vector (dx, dy, dz) defining its axis and the two radii r1 and r2 of the faces (these radii can be zero). If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. angle defines the optional angular opening (from 0 to 2*Pi). Return the tag of the cone.

Return an integer value.

gmsh.model.occ.addWedge — Function

gmsh.model.occ.addWedge(x, y, z, dx, dy, dz, tag = -1, ltx = 0.)

Add a right angular wedge in the OpenCASCADE CAD representation, defined by the right-angle point (x, y, z) and the 3 extends along the x-, y- and z-axes (dx, dy, dz). If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. The optional argument ltx defines the top extent along the x-axis. Return the tag of the wedge.

Return an integer value.

gmsh.model.occ.addTorus — Function

gmsh.model.occ.addTorus(x, y, z, r1, r2, tag = -1, angle = 2*pi)

Add a torus in the OpenCASCADE CAD representation, defined by its center (x, y, z) and its 2 radii r and r2. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. The optional argument angle defines the angular opening (from 0 to 2*Pi). Return the tag of the wedge.

Return an integer value.

gmsh.model.occ.addThruSections — Function

gmsh.model.occ.addThruSections(wireTags, tag = -1, makeSolid = true, makeRuled = false, maxDegree = -1)

Add a volume (if the optional argument makeSolid is set) or surfaces in the OpenCASCADE CAD representation, defined through the open or closed wires wireTags. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically. The new entities are returned in outDimTags. If the optional argument makeRuled is set, the surfaces created on the boundary are forced to be ruled surfaces. If maxDegree is positive, set the maximal degree of resulting surface.

Return outDimTags.

gmsh.model.occ.addThickSolid — Function

gmsh.model.occ.addThickSolid(volumeTag, excludeSurfaceTags, offset, tag = -1)

Add a hollowed volume in the OpenCASCADE CAD representation, built from an initial volume volumeTag and a set of faces from this volume excludeSurfaceTags, which are to be removed. The remaining faces of the volume become the walls of the hollowed solid, with thickness offset. If tag is positive, set the tag explicitly; otherwise a new tag is selected automatically.

Return outDimTags.

gmsh.model.occ.extrude — Function

gmsh.model.occ.extrude(dimTags, dx, dy, dz, numElements = Cint[], heights = Cdouble[], recombine = false)

Extrude the entities dimTags in the OpenCASCADE CAD representation, using a translation along (dx, dy, dz). Return extruded entities in outDimTags. If numElements is not empty, also extrude the mesh: the entries in numElements give the number of elements in each layer. If height is not empty, it provides the (cumulative) height of the different layers, normalized to 1. If recombine is set, recombine the mesh in the layers.

Return outDimTags.

gmsh.model.occ.revolve — Function

gmsh.model.occ.revolve(dimTags, x, y, z, ax, ay, az, angle, numElements = Cint[], heights = Cdouble[], recombine = false)

Extrude the entities dimTags in the OpenCASCADE CAD representation, using a rotation of angle radians around the axis of revolution defined by the point (x, y, z) and the direction (ax, ay, az). Return extruded entities in outDimTags. If numElements is not empty, also extrude the mesh: the entries in numElements give the number of elements in each layer. If height is not empty, it provides the (cumulative) height of the different layers, normalized to 1. When the mesh is extruded the angle should be strictly smaller than 2*Pi. If recombine is set, recombine the mesh in the layers.

Return outDimTags.

gmsh.model.occ.addPipe — Function

gmsh.model.occ.addPipe(dimTags, wireTag)

Add a pipe in the OpenCASCADE CAD representation, by extruding the entities dimTags along the wire wireTag. Return the pipe in outDimTags.

Return outDimTags.

gmsh.model.occ.fillet — Function

gmsh.model.occ.fillet(volumeTags, curveTags, radii, removeVolume = true)

Fillet the volumes volumeTags on the curves curveTags with radii radii. The radii vector can either contain a single radius, as many radii as curveTags, or twice as many as curveTags (in which case different radii are provided for the begin and end points of the curves). Return the filleted entities in outDimTags. Remove the original volume if removeVolume is set.

Return outDimTags.

gmsh.model.occ.chamfer — Function

gmsh.model.occ.chamfer(volumeTags, curveTags, surfaceTags, distances, removeVolume = true)

Chamfer the volumes volumeTags on the curves curveTags with distances distances measured on surfaces surfaceTags. The distances vector can either contain a single distance, as many distances as curveTags and surfaceTags, or twice as many as curveTags and surfaceTags (in which case the first in each pair is measured on the corresponding surface in surfaceTags, the other on the other adjacent surface). Return the chamfered entities in outDimTags. Remove the original volume if removeVolume is set.

Return outDimTags.

gmsh.model.occ.fuse — Function

gmsh.model.occ.fuse(objectDimTags, toolDimTags, tag = -1, removeObject = true, removeTool = true)

Compute the boolean union (the fusion) of the entities objectDimTags and toolDimTags in the OpenCASCADE CAD representation. Return the resulting entities in outDimTags. If tag is positive, try to set the tag explicitly (only valid if the boolean operation results in a single entity). Remove the object if removeObject is set. Remove the tool if removeTool is set.

Return outDimTags, outDimTagsMap.

gmsh.model.occ.intersect — Function

gmsh.model.occ.intersect(objectDimTags, toolDimTags, tag = -1, removeObject = true, removeTool = true)

Compute the boolean intersection (the common parts) of the entities objectDimTags and toolDimTags in the OpenCASCADE CAD representation. Return the resulting entities in outDimTags. If tag is positive, try to set the tag explicitly (only valid if the boolean operation results in a single entity). Remove the object if removeObject is set. Remove the tool if removeTool is set.

Return outDimTags, outDimTagsMap.

gmsh.model.occ.cut — Function

gmsh.model.occ.cut(objectDimTags, toolDimTags, tag = -1, removeObject = true, removeTool = true)

Compute the boolean difference between the entities objectDimTags and toolDimTags in the OpenCASCADE CAD representation. Return the resulting entities in outDimTags. If tag is positive, try to set the tag explicitly (only valid if the boolean operation results in a single entity). Remove the object if removeObject is set. Remove the tool if removeTool is set.

Return outDimTags, outDimTagsMap.

gmsh.model.occ.fragment — Function

gmsh.model.occ.fragment(objectDimTags, toolDimTags, tag = -1, removeObject = true, removeTool = true)

Compute the boolean fragments (general fuse) of the entities objectDimTags and toolDimTags in the OpenCASCADE CAD representation. Return the resulting entities in outDimTags. If tag is positive, try to set the tag explicitly (only valid if the boolean operation results in a single entity). Remove the object if removeObject is set. Remove the tool if removeTool is set.

Return outDimTags, outDimTagsMap.

gmsh.model.occ.translate — Function

gmsh.model.occ.translate(dimTags, dx, dy, dz)

Translate the entities dimTags in the OpenCASCADE CAD representation along (dx, dy, dz).

gmsh.model.occ.rotate — Function

gmsh.model.occ.rotate(dimTags, x, y, z, ax, ay, az, angle)

Rotate the entities dimTags in the OpenCASCADE CAD representation by angle radians around the axis of revolution defined by the point (x, y, z) and the direction (ax, ay, az).

gmsh.model.occ.dilate — Function

gmsh.model.occ.dilate(dimTags, x, y, z, a, b, c)

Scale the entities dimTags in the OpenCASCADE CAD representation by factors a, b and c along the three coordinate axes; use (x, y, z) as the center of the homothetic transformation.

gmsh.model.occ.mirror — Function

gmsh.model.occ.mirror(dimTags, a, b, c, d)

Mirror the entities dimTags in the OpenCASCADE CAD representation, with respect to the plane of equation a * x + b * y + c * z + d = 0.

gmsh.model.occ.symmetrize — Function

gmsh.model.occ.symmetrize(dimTags, a, b, c, d)

Mirror the entities dimTags in the OpenCASCADE CAD representation, with respect to the plane of equation a * x + b * y + c * z + d = 0. (This is a synonym for mirror, which will be deprecated in a future release.)

gmsh.model.occ.affineTransform — Function

gmsh.model.occ.affineTransform(dimTags, a)

Apply a general affine transformation matrix a (16 entries of a 4x4 matrix, by row; only the 12 first can be provided for convenience) to the entities dimTags in the OpenCASCADE CAD representation.

gmsh.model.occ.copy — Function

gmsh.model.occ.copy(dimTags)

Copy the entities dimTags in the OpenCASCADE CAD representation; the new entities are returned in outDimTags.

Return outDimTags.

gmsh.model.occ.remove — Function

gmsh.model.occ.remove(dimTags, recursive = false)

Remove the entities dimTags in the OpenCASCADE CAD representation. If recursive is true, remove all the entities on their boundaries, down to dimension 0.

gmsh.model.occ.removeAllDuplicates — Function

gmsh.model.occ.removeAllDuplicates()

Remove all duplicate entities in the OpenCASCADE CAD representation (different entities at the same geometrical location) after intersecting (using boolean fragments) all highest dimensional entities.

gmsh.model.occ.healShapes — Function

gmsh.model.occ.healShapes(dimTags = Tuple{Cint,Cint}[], tolerance = 1e-8, fixDegenerated = true, fixSmallEdges = true, fixSmallFaces = true, sewFaces = true, makeSolids = true)

Apply various healing procedures to the entities dimTags (or to all the entities in the model if dimTags is empty) in the OpenCASCADE CAD representation. Return the healed entities in outDimTags. Available healing options are listed in the Gmsh reference manual.

Return outDimTags.

gmsh.model.occ.importShapes — Function

gmsh.model.occ.importShapes(fileName, highestDimOnly = true, format = "")

Import BREP, STEP or IGES shapes from the file fileName in the OpenCASCADE CAD representation. The imported entities are returned in outDimTags. If the optional argument highestDimOnly is set, only import the highest dimensional entities in the file. The optional argument format can be used to force the format of the file (currently "brep", "step" or "iges").

Return outDimTags.

gmsh.model.occ.getEntities — Function

gmsh.model.occ.getEntities(dim = -1)

Get all the OpenCASCADE entities. If dim is >= 0, return only the entities of the specified dimension (e.g. points if dim == 0). The entities are returned as a vector of (dim, tag) integer pairs.

Return dimTags.

gmsh.model.occ.getEntitiesInBoundingBox — Function

gmsh.model.occ.getEntitiesInBoundingBox(xmin, ymin, zmin, xmax, ymax, zmax, dim = -1)

Get the OpenCASCADE entities in the bounding box defined by the two points (xmin, ymin, zmin) and (xmax, ymax, zmax). If dim is >= 0, return only the entities of the specified dimension (e.g. points if dim == 0).

Return tags.

gmsh.model.occ.getBoundingBox — Function

gmsh.model.occ.getBoundingBox(dim, tag)

Get the bounding box (xmin, ymin, zmin), (xmax, ymax, zmax) of the OpenCASCADE entity of dimension dim and tag tag.

Return xmin, ymin, zmin, xmax, ymax, zmax.

gmsh.model.occ.getMass — Function

gmsh.model.occ.getMass(dim, tag)

Get the mass of the OpenCASCADE entity of dimension dim and tag tag.

Return mass.

gmsh.model.occ.getCenterOfMass — Function

gmsh.model.occ.getCenterOfMass(dim, tag)

Get the center of mass of the OpenCASCADE entity of dimension dim and tag tag.

Return x, y, z.

gmsh.model.occ.getMatrixOfInertia — Function

gmsh.model.occ.getMatrixOfInertia(dim, tag)

Get the matrix of inertia (by row) of the OpenCASCADE entity of dimension dim and tag tag.

Return mat.

gmsh.model.occ.getMaxTag — Function

gmsh.model.occ.getMaxTag(dim)

Get the maximum tag of entities of dimension dim in the OpenCASCADE CAD representation.

Return an integer value.

gmsh.model.occ.setMaxTag — Function

gmsh.model.occ.setMaxTag(dim, maxTag)

Set the maximum tag maxTag for entities of dimension dim in the OpenCASCADE CAD representation.

gmsh.model.occ.synchronize — Function

gmsh.model.occ.synchronize()

Synchronize the OpenCASCADE CAD representation with the current Gmsh model. This can be called at any time, but since it involves a non trivial amount of processing, the number of synchronization points should normally be minimized. Without synchronization the entities in the OpenCASCADE CAD representation are not available to any function outside of the OpenCASCADE CAD kernel functions.

Meshing Constraints with OCC Kernel

gmsh.model.occ.mesh.setSize — Function

gmsh.model.occ.mesh.setSize(dimTags, size)

Set a mesh size constraint on the entities dimTags in the OpenCASCADE CAD representation. Currently only entities of dimension 0 (points) are handled.