Gmsh Options

gmsh.option.setNumber — Function

gmsh.option.setNumber(name, value)

Set a numerical option to value. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual.

gmsh.option.getNumber — Function

gmsh.option.getNumber(name)

Get the value of a numerical option. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual.

Return value.

gmsh.option.setString — Function

gmsh.option.setString(name, value)

Set a string option to value. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual.

gmsh.option.getString — Function

gmsh.option.getString(name)

Get the value of a string option. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual.

Return value.

gmsh.option.setColor — Function

gmsh.option.setColor(name, r, g, b, a = 255)

Set a color option to the RGBA value (r, g, b, a), where where r, g, b and a should be integers between 0 and 255. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual, with the "Color." middle string removed.

gmsh.option.getColor — Function

gmsh.option.getColor(name)

Get the r, g, b, a value of a color option. name is of the form "category.option" or "category[num].option". Available categories and options are listed in the Gmsh reference manual, with the "Color." middle string removed.

Return r, g, b, a.