gmsh.jl

gmsh.jl wraps the Gmsh binaries (gmsh_jll.jl) and API in one registered package. The registered package just packs and gathers already built software. Issues related to Gmsh needs to be reported here.

Top Level Functions

The developers of gmsh decided to split all functionalities across modules. However, a few top level functions exists:

gmsh.initialize — Function

gmsh.initialize(argv = Vector{String}(), readConfigFiles = true)

Initialize Gmsh API. This must be called before any call to the other functions in the API. If argc and argv (or just argv in Python or Julia) are provided, they will be handled in the same way as the command line arguments in the Gmsh app. If readConfigFiles is set, read system Gmsh configuration files (gmshrc and gmsh-options). Initializing the API sets the options "General.Terminal" to 1 and "General.AbortOnError" to 2.

gmsh.finalize — Function

gmsh.finalize()

Finalize the Gmsh API. This must be called when you are done using the Gmsh API.

gmsh.open — Function

gmsh.open(fileName)

Open a file. Equivalent to the File->Open menu in the Gmsh app. Handling of the file depends on its extension and/or its contents: opening a file with model data will create a new model.

gmsh.merge — Function

gmsh.merge(fileName)

Merge a file. Equivalent to the File->Merge menu in the Gmsh app. Handling of the file depends on its extension and/or its contents. Merging a file with model data will add the data to the current model.

gmsh.write — Function

gmsh.write(fileName)

Write a file. The export format is determined by the file extension.

gmsh.clear — Function

gmsh.clear()

Clear all loaded models and post-processing data, and add a new empty model.